Use Int instead of Integer to configure #partitions and replication factor

+++ src/test/scala/com/miguno/kafkastorm/spark/KafkaSparkStreamingSpec.scala
@@ -171,7 +171,7 @@ class KafkaSparkStreamingSpec extends FeatureSpec with Matchers with BeforeAndAf
         // And yes, the way we do this looks quite strange -- we combine a hardcoded `1` in the topic map with a
         // subsequent `(1..N)` construct.  But this approach is the recommended way.
         val topics = Map(inputTopic.name -> 1)
-        val streams = (1 to inputTopic.partitions.toInt) map { _ =>
+        val streams = (1 to inputTopic.partitions) map { _ =>
           KafkaUtils.createStream[Array[Byte], Array[Byte], DefaultDecoder, DefaultDecoder](
             ssc,
             kafkaParams,
